DRAKEN-4295: Fixed sonar issues
